requirements.txt
Flask
PyGitHub
requests
python-dotenv
google-generativeai
docker
python-magic
